Move signing function to wallet struct
